io: make fairness_counter size_t, not int.
io_loop() increments the static fairness_counter on every iteration;
after 2^31 iterations signed overflow is undefined behavior (the
rotation index math it feeds is harmless on real ABIs, but the UB
poisons LTO/sanitizer-clean builds).  size_t wrap is well-defined.

Found by audit (see audit-findings/io.md F3).

+++ ccan/io/poll.c
@@ -386,7 +386,7 @@ void *io_loop(struct timers *timers, struct timer **expired)
 {
 	void *ret;
 	/* This ensures we don't always service lower fds first */
-	static int fairness_counter;
+	static size_t fairness_counter;
 
 	/* if timers is NULL, expired must be.  If not, not. */
 	assert(!timers == !expired);
